Optimization: SOLID, Single Responsibility
The class broke the single responsability  design principles of SOLID. I was using the player class to store a turn boolean, to keep track of which players turn was next. I decided to implement my own custom circular arraylist, that keeps track of which player is next, as well as changing the end pointer when someone raises. Now the class is much more clear and concise.