Add files via upload
To keep track of players raising and calling, you have to think about them in a circular array, so that when player C raises, player B becomes the end of the array. My program was working without this class, but it clustered my Player and PlayerPanel classes with unnecessary code. This is a better implementation.